refactor(HebrewLettersStatsPanel): extract StatPanelCard, LetterSpecificStats, PracticeHistoryList (#144)
